requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
ansicolors==1.1.8
ansiwrap==0.8.4
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
bokeh==2.4.1
boto3==1.26.30
botocore==1.29.30
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
chainmap==1.0.3
charset-normalizer==2.0.12
click==8.1.3
combomethod==1.0.12
commonmark==0.9.1
datasets==2.3.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
happytransformer==2.4.1
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
language-tool-python==2.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mementos==1.3.1
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
nulltype==2.3.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
options==1.4.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pox==0.3.1
ppft==1.7.6.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-transformers==1.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.1.0
regex==2022.6.2
requests==2.28.0
responses==0.18.0
rich==12.4.4
s3transfer==0.6.0
sacremoses==0.0.53
say==1.6.6
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
simplere==1.2.13
six==1.12.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
st-clickable-images==0.0.3
stack-data==0.3.0
starlette==0.19.1
streamlit==1.10.0
streamlit-bokeh-events==0.1.2
streamlit-chat==0.0.2.1
terminado==0.15.0
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.11.2
torch==1.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.18.1
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
